requirements.txt
# ============================================================
# Real-Time Face Mask & Emotion Detection System
# Requirements File
# ============================================================

# Core Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Computer Vision
opencv-python>=4.7.0
opencv-contrib-python>=4.7.0

# Image Processing
Pillow>=9.3.0
numpy>=1.23.0
imutils>=0.5.3

# Data Science & Visualization
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
pandas>=1.5.0

# Web Framework (Flask)
Flask>=2.2.0
Flask-SocketIO>=5.3.0

# Streamlit App
streamlit>=1.20.0

# Audio Alert
playsound>=1.3.0
pygame>=2.1.0

# Logging & Utilities
loguru>=0.6.0
tqdm>=4.64.0
colorama>=0.4.6

# API
flask-cors>=3.0.10

# Optional: GPU Support
# tensorflow-gpu>=2.10.0  # Uncomment if CUDA GPU available

# Model Download / Dataset
requests>=2.28.0
kaggle>=1.5.12

# Dockerfile utilities
gunicorn>=20.1.0